Ship npm integration test dependencies in dist tarball
This avoids having to run `npm`, and thus a potentially brittle internet
access, during FMF test runs. It also avoids the npm test dependency,
which tends to be uninstallable in Rawhide.

We have successfully used this approach in cockpit for a long time.

+++ test/browser/run-test.sh
@@ -3,21 +3,12 @@ set -eux
 
 # tests need cockpit's bots/ libraries and test infrastructure
 cd $SOURCE
-git init
 rm -f bots  # common local case: existing bots symlink
 make bots
 
-# support running from clean git tree
-if [ ! -d node_modules/chrome-remote-interface ]; then
-    # copy package.json temporarily otherwise npm might try to install the dependencies from it
-    mv package.json .package.json
-    npm install chrome-remote-interface sizzle
-    mv .package.json package.json
-fi
-
 # disable detection of affected tests; testing takes too long as there is no parallelization,
 # and TF machines are slow and brittle
-mv .git dot-git
+[ ! -e .git ] || mv .git dot-git
 
 . /usr/lib/os-release
 export TEST_OS="${ID}-${VERSION_ID/./-}"
